Reorganiza file structure
