+++ src/Components/App/App.test.js
@@ -1,0 +1,2 @@
+import React from 'react';
+import App from './App';
